% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgr5_gfp_mouse_cellTypes.R
\docType{data}
\name{lgr5_gfp_mouse_cellTypes.newCelltypes}
\alias{lgr5_gfp_mouse_cellTypes.newCelltypes}
\title{lgr5_gfp_mouse_cellTypes.newCelltypes}
\format{Tibble with:
\describe{
\item{Samples}{Sample names}
\item{class}{Cell type}
}}
\usage{
lgr5_gfp_mouse_cellTypes.newCelltypes
}
\description{
lgr5 gfp mouse cellTypes
}
\details{
Description.
}
\examples{
data(lgr5_gfp_mouse_cellTypes.newCelltypes)
}
\keyword{datasets}
