% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgr5_gfp_mouse_cellTypes.R
\docType{data}
\name{lgr5_gfp_mouse_cellTypes.uObj}
\alias{lgr5_gfp_mouse_cellTypes.uObj}
\title{lgr5_gfp_mouse_cellTypes.uObj}
\format{spUnsupervised object.}
\usage{
lgr5_gfp_mouse_cellTypes.uObj
}
\description{
lgr5 gfp mouse cellTypes
}
\details{
Description.
}
\examples{
data(lgr5_gfp_mouse_cellTypes.uObj)
}
\keyword{datasets}
